remove bracket
