+++ RCA_snejana/main.cpp
@@ -1,0 +1,11 @@
+#include <QCoreApplication>
+#include "rca.h"
+
+int main(int argc, char *argv[])
+{
+    QCoreApplication a(argc, argv);
+
+    Rca rca;
+
+    return a.exec();
+}
