--- main.cpp
@@ -1,11 +1,0 @@
-#include <QCoreApplication>
-#include "rca.h"
-
-int main(int argc, char *argv[])
-{
-    QCoreApplication a(argc, argv);
-
-    Rca rca;
-
-    return a.exec();
-}
